Move WordType declarations to a separate file
